refactor(routines): extract UI components and modularize routine screens
This commit refactors the routines feature by moving locally defined composables into a dedicated `components` package. This improves code maintainability and reusability across the routine detail and editor screens.

- **`ui/components/`**:
    - Created several new files (`DailyInsightSmall.kt`, `EditStepForm.kt`, `SelectionPages.kt`, `SplitRitualStep.kt`, `StepHeroPage.kt`, `StepSummaryRow.kt`) to house modular UI elements.
    - **`RoutineComponents.kt`**: Consolidated primary ritual cards and banners, such as `HeroRitualCard` and `DailyInsightBanner`, into a shared file.
    - **`SelectionPages.kt`**: Implemented specialized pages for `Macro`, `Micro`, and `Item` selection to support the product attachment flow.
- **`RoutineDetailScreen.kt`, `RoutineEditorScreen.kt` & `RoutinesScreen.kt`**:
    - Significantly reduced file complexity by removing inline composable definitions and migrating to the new shared components.
    - Updated `RoutineDetailScreen` to use the refactored `SplitRitualStep` and `DailyInsightSmall` components.
    - Enhanced `RoutineEditorScreen` with a standard `RoutineEditorScreenPreview` and integrated new form components.
- **`strings.xml`**:
    - Added a new string resource for the "Save" action to support localization and consistency in the editor UI.
