+++ applications/kocolor/features/routines/src/main/res/values/strings.xml
@@ -16,4 +16,5 @@
     <string name="applications_kocolor_features_routines_daily_insight">Daily Insight</string>
     <string name="applications_kocolor_features_routines_insight_quote">\"Patience is the foundation of every lasting ritual.\"</string>
     <string name="applications_kocolor_features_routines_view_progress">VIEW PROGRESS</string>
+    <string name="applications_kocolor_features_routines_save">Save</string>
 </resources>
